tweak padding & weight
Trying to get the layout to respond as well as it can to resizing since the AppWidgetProvider can't help (unless it watches for screen rotations) An alternative is to use different resources for portrait mode.

The aim is to prioritize seeing appwidget_text2 over appwidget_text1 on a variety of screen sizes, densities, and orientations.